flujo bd y check

+++ src/main/java/org/lemon/core/usecase/interfaces/ProcessMessage.java
@@ -3,5 +3,5 @@ package org.lemon.core.usecase.interfaces;
 import org.lemon.core.entity.FuckOff;
 
 public interface ProcessMessage {
-    FuckOff execute(int userId);
+    FuckOff execute(Long userId);
 }
